+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+### v1.1.0
+- Fixed modded commands with spaces or other punctuation not working properly
+- The new "Remove Command Punctuation" config option can be disabled incase this causes conflicts with any terminal mods
+
 ### v1.0.0
 - Initial release